aws-monitoring/requirements.txt
flask==3.0.0
boto3==1.34.0
pyyaml==6.0.1
requests==2.31.0
botocore==1.34.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
moto==5.0.0
pytest-flask==1.3.0
